- Plugin: Prevent multiple build runs at same time with UI feedback. - Plugin: Update regex validator for path on windows to be less restrictive. - Plugin: Remove un-needed calls to DUP_PRO variables. - Plugin: Domain check on step 1 installer db host less restrictive. - Plugin: Improved file permission changes to restore to original permission sets. - Plugin: Root directory is no longer selected from the quick filters on step 2 scanner. - Plugin: Quick Filter checkbox notice message when checked but not applied. - Plugin: Removed 'FilesystemIterator' class to still allow support for PHP 5.2. - Installer: Fixed code reference for setCharset referencing bad variable call. - Installer: Auto disable "Really Simple SSL" and "Simple Google reCAPTCHA" at install time to avoid login issues. - Installer: Cookie domain should be restricted to installer domain and path. - Installer: General Multisite cleanup.
git-svn-id: https://plugins.svn.wordpress.org/duplicator/trunk@2000399 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ views/packages/main/packages.php
@@ -4,6 +4,7 @@
     $packages = DUP_Package::get_all();
 	$totalElements	= count($packages);
 	$statusCount	= 0; // total packages completed
+    $active_package_present = DUP_Package::is_active_package_present();
 
     $package_debug	= DUP_Settings::Get('package_debug');
     $ajax_nonce		= wp_create_nonce('package_list');
@@ -23,16 +24,27 @@
 	input#dup-bulk-action-all {margin:0px;padding:0px 0px 0px 5px;}
 	button.dup-button-selected {border:1px solid #000 !important; background-color:#dfdfdf !important;}
 	div.dup-quick-start {font-style:italic; font-size: 13px; line-height: 18px; margin-top: 15px}
+
+    .add-new-h2.disabled { 
+        cursor: not-allowed;
+        border-color: #ccc !important;
+        background: #f7f7f7 !important;
+        color: #ccc !important;
+    }
 	
 	/* Table package details */
 	table.dup-pack-table {word-break:break-all;}
 	table.dup-pack-table th {white-space:nowrap !important;}
 	table.dup-pack-table td.pack-name {text-overflow:ellipsis; white-space:nowrap}
+    table.dup-pack-table td.pack-size {min-width: 65px; }
 
 	table.dup-pack-table input[name="delete_confirm"] {margin-left:15px}
 	table.dup-pack-table td.fail {border-left: 4px solid #d54e21;}
 	table.dup-pack-table td.pass {border-left: 4px solid #2ea2cc;}
-	tr.dup-pack-info td {white-space:nowrap; padding:12px 30px 0px 7px;}
+
+    .dup-pack-info {height: 45px;}
+    .dup-pack-info td {vertical-align: middle; }
+	tr.dup-pack-info td {white-space:nowrap; padding:2px 30px 2px 7px;}
 	tr.dup-pack-info td.get-btns {text-align:right; padding:3px 5px 6px 0px !important;}
 	tr.dup-pack-info td.get-btns button {box-shadow:none}
 	textarea.dup-pack-debug {width:98%; height:300px; font-size:11px; display:none}
@@ -42,6 +54,13 @@
 	div#dup-help-dlg i {display: inline-block; width: 15px; padding:2px;line-height:28px; font-size:14px;}
 	tr.dup-pack-info sup  {font-style:italic;font-size:10px; cursor: pointer; vertical-align: baseline; position: relative; top: -0.8em;}
 	tr#pack-processing {display: none}
+    
+    /* Building package */
+ 
+    .dup-pack-info .building-info {display: none;}
+    .dup-pack-info.is-running .building-info {display: inline;}
+    .dup-pack-info.is-running .get-btns button {display: none;}
+
 </style>
 
 <form id="form-duplicator" method="post">
@@ -68,11 +87,16 @@ TOOL-BAR -->
 			$package_url = admin_url('admin.php?page=duplicator&tab=new1');
 			$package_nonce_url = wp_nonce_url($package_url, 'new1-package');
 			?>
-			<a href="<?php echo $package_nonce_url;?>" class="add-new-h2"><?php esc_html_e("Create New", 'duplicator'); ?></a>
+			<a id="dup-create-new" 
+               onClick="return Duplicator.Pack.CreateNew(this);"
+               href="<?php echo $package_nonce_url;?>"
+               class="add-new-h2 <?php echo ($active_package_present ? 'disabled' : ''); ?>"
+               >
+                <?php esc_html_e("Create New", 'duplicator'); ?>
+            </a>
 		</td>
 	</tr>
-</table>	
-
+</table>
 
 <?php if($totalElements == 0 )  : ?>
 	<!-- ====================
@@ -137,31 +161,36 @@ TOOL-BAR -->
 		$txt_mode_zip  = __('Archive created as zip file', 'duplicator');
 		$txt_mode_daf  = __('Archive created as daf file', 'duplicator');
 		//$rows = $qryResult;
+
+
 		foreach ($packages as $Package) {
-            
+            $is_running_package = $Package->isRunning();
+
             // Never display incomplete packages and purge those that are no longer active
-            if($Package->Status >= 0 && $Package->Status < 100) {
-                if(DUP_Settings::Get('active_package_id') != $Package->ID) {
-                    $Package->delete();
+            if(!$is_running_package && $Package->Status >= 0 && $Package->Status < 100) {
+                $Package->delete();
+                
+                if ($rowCount <= 1 && $totalElements == 1) {
+                    $package_running = true;
                 }
 
-				if ($rowCount <= 1 && $totalElements == 1)
-					$package_running = true;
-				
-				continue;
+                continue;
             }
+            
             
 			$pack_dbonly = false;
 
 			if (is_object($Package)) {
 				 $pack_name			= $Package->Name;
-				 $pack_archive_size = $Package->Archive->Size;
+				 $pack_archive_size = $Package->getArchiveSize();
+                 $pack_perc         = $Package->Status;
 				 $pack_storeurl		= $Package->StoreURL;
 				 $pack_namehash	    = $Package->NameHash;
 				 $pack_dbonly       = $Package->Archive->ExportOnlyDB;
 				 $pack_build_mode   = ($Package->Archive->Format === 'ZIP') ? true : false;
 			} else {
 				 $pack_archive_size = 0;
+                 $pack_perc         = 0;
 				 $pack_storeurl		= 'unknown';
 				 $pack_name			= 'unknown';
 				 $pack_namehash	    = 'unknown';
@@ -183,10 +212,10 @@ TOOL-BAR -->
 
 			<?php
 
-            if ($Package->Status >= 100) :
+            if ($Package->Status >= 100 || $is_running_package) :
                 $statusCount ++;
                 ?>
-				<tr class="dup-pack-info <?php echo esc_attr($css_alt); ?>">
+				<tr class="dup-pack-info <?php echo esc_attr($css_alt); ?> <?php echo $is_running_package ? 'is-running' : ''; ?>">
 					<td class="pass"><input name="delete_confirm" type="checkbox" id="<?php echo absint($Package->ID); ?>" /></td>
 					<td>
 						<?php 
@@ -194,9 +223,10 @@ TOOL-BAR -->
 							echo ($pack_build_mode) ? " <sup title='{$txt_mode_zip}'>zip</sup>" : " <sup title='{$txt_mode_daf}'>daf</sup>";
 						?>
 					</td>
-					<td><?php echo DUP_Util::byteSize($pack_archive_size); ?></td>
+					<td class="pack-size"><?php echo DUP_Util::byteSize($pack_archive_size); ?></td>
 					<td class='pack-name'>
 						<?php	echo ($pack_dbonly) ? "{$pack_name} <sup title='".esc_attr($txt_dbonly)."'>DB</sup>" : esc_html($pack_name); ?>
+                        <span class="building-info" ><i class="fa fa-gear fa-spin"></i> <b>Building Package</b> <span class="perc"><?php echo $pack_perc; ?></span>%</span>
 					</td>
 					<td class="get-btns">
 						<button id="<?php echo esc_attr("{$uniqueid}_installer.php"); ?>" class="button no-select" onclick="Duplicator.Pack.DownloadPackageFile(0, <?php echo absint($Package->ID); ?>); return false;">
@@ -215,19 +245,19 @@ TOOL-BAR -->
 			<?php else : ?>	
 			
 				<?php
-					$size = 0;
+					/*$size = 0;
 					$tmpSearch = glob(DUPLICATOR_SSDIR_PATH_TMP . "/{$pack_namehash}_*");
 					if (is_array($tmpSearch)) {
 						$result = array_map('filesize', $tmpSearch);
 						$size = array_sum($result);
 					}
-					$pack_archive_size = $size;
+					$pack_archive_size = $size;*/
 					$error_url = "?page=duplicator&action=detail&tab=detail&id={$Package->ID}";
 				?>
 				<tr class="dup-pack-info  <?php echo esc_attr($css_alt); ?>">
 					<td class="fail"><input name="delete_confirm" type="checkbox" id="<?php echo absint($Package->ID); ?>" /></td>
 					<td><?php echo DUP_Package::getCreatedDateFormat($Package->Created, $ui_create_frmt);?></td>
-					<td><?php echo DUP_Util::byteSize($size); ?></td>
+					<td class="pack-size"><?php echo DUP_Util::byteSize($pack_archive_size); ?></td>
 					<td class='pack-name'><?php echo esc_html($pack_name); ?></td>               
 					<td class="get-btns error-msg" colspan="2">		
 						<span>
@@ -270,31 +300,36 @@ TOOL-BAR -->
 <!-- ==========================================
 THICK-BOX DIALOGS: -->
 <?php
-	$alert1 = new DUP_UI_Dialog();
-	$alert1->title		= __('Bulk Action Required', 'duplicator');
-	$alert1->message	= '<i class="fa fa-exclamation-triangle"></i>&nbsp;';
-	$alert1->message	.= __('No selections made! Please select an action from the "Bulk Actions" drop down menu.', 'duplicator');
-	$alert1->initAlert();
-	
-	$alert2 = new DUP_UI_Dialog();
-	$alert2->title		= __('Selection Required', 'duplicator', 'duplicator');
-	$alert2->message	= '<i class="fa fa-exclamation-triangle"></i>&nbsp;';
-	$alert2->message	.= __('No selections made! Please select at least one package to delete.', 'duplicator');
-	$alert2->initAlert();
-	
-	$confirm1 = new DUP_UI_Dialog();
-	$confirm1->title			= __('Delete Packages?', 'duplicator');
-	$confirm1->message			= __('Are you sure, you want to delete the selected package(s)?', 'duplicator');
-	$confirm1->progressText	= __('Removing Packages, Please Wait...', 'duplicator');
-	$confirm1->jscallback		= 'Duplicator.Pack.Delete()';
-	$confirm1->initConfirm();
+$alert1          = new DUP_UI_Dialog();
+$alert1->title   = __('Bulk Action Required', 'duplicator');
+$alert1->message = '<i class="fa fa-exclamation-triangle"></i>&nbsp;';
+$alert1->message .= __('No selections made! Please select an action from the "Bulk Actions" drop down menu.', 'duplicator');
+$alert1->initAlert();
 
-	$alert3 = new DUP_UI_Dialog();
-	$alert3->height     = 355;
-	$alert3->width      = 350;
-	$alert3->title		= __('Duplicator Help', 'duplicator');
-	$alert3->message	= "<div id='dup-help-dlg'></div>";
-	$alert3->initAlert();
+$alert2          = new DUP_UI_Dialog();
+$alert2->title   = __('Selection Required', 'duplicator', 'duplicator');
+$alert2->message = '<i class="fa fa-exclamation-triangle"></i>&nbsp;';
+$alert2->message .= __('No selections made! Please select at least one package to delete.', 'duplicator');
+$alert2->initAlert();
+
+$confirm1               = new DUP_UI_Dialog();
+$confirm1->title        = __('Delete Packages?', 'duplicator');
+$confirm1->message      = __('Are you sure, you want to delete the selected package(s)?', 'duplicator');
+$confirm1->progressText = __('Removing Packages, Please Wait...', 'duplicator');
+$confirm1->jscallback   = 'Duplicator.Pack.Delete()';
+$confirm1->initConfirm();
+
+$alert3          = new DUP_UI_Dialog();
+$alert3->height  = 355;
+$alert3->width   = 350;
+$alert3->title   = __('Duplicator Help', 'duplicator');
+$alert3->message = "<div id='dup-help-dlg'></div>";
+$alert3->initAlert();
+
+$alertPackRunning          = new DUP_UI_Dialog();
+$alertPackRunning->title   = __('Alert!', 'duplicator');
+$alertPackRunning->message = __('A package is being processed. Retry later.', 'duplicator');
+$alertPackRunning->initAlert();
 ?>
 
 <!-- =======================
@@ -317,7 +352,26 @@ DIALOG: HELP DIALOG -->
 <script>
 jQuery(document).ready(function($) 
 {
-	
+    /** Create new package check */
+    Duplicator.Pack.CreateNew = function(e){
+        var cButton = $(e);
+        if (cButton.hasClass('disabled')) {
+            <?php $alertPackRunning->showAlert(); ?>
+        } else {
+            Duplicator.Pack.GetActivePackageInfo(function (info) {
+                if (info.present) {
+                    cButton.addClass('disabled');
+                    // reloag current page to update packages list
+                    location.reload(true);
+                } else {
+                    // no active package. Load step1 page.
+                    window.location = cButton.attr('href');
+                }
+            });
+        }
+        return false;
+    };
+
 	/*	Creats a comma seperate list of all selected package ids  */
 	Duplicator.Pack.GetDeleteList = function () 
 	{
@@ -365,6 +419,55 @@ jQuery(document).ready(function($)
 		});
 
 	};
+
+    Duplicator.Pack.ActivePackageInfo = function (info) {
+        $('.dup-pack-info.is-running .pack-size').text(info.size_format);
+
+        if (info.present) {
+            $('.dup-pack-info.is-running .building-info .perc').text(info.status);
+
+            setTimeout(function(){
+                Duplicator.Pack.GetActivePackageInfo(Duplicator.Pack.ActivePackageInfo);
+            }, 1000);
+            
+        } else {
+            $('.dup-pack-info.is-running').removeClass('is-running');
+            $('#dup-create-new.disabled').removeClass('disabled');
+        }
+    }
+
+    /*	Get active package info
+	 *
+     *	  */
+	Duplicator.Pack.GetActivePackageInfo = function (callbackOnSuccess)
+	{
+		$.ajax({
+            type: "POST",
+            cache: false,
+            url: ajaxurl,
+            dataType: "json",
+            timeout: 10000000,
+			data: {
+                action : 'duplicator_active_package_info',
+                nonce: '<?php echo esc_js(wp_create_nonce('duplicator_active_package_info')); ?>'
+            },
+			complete: function () {},
+            success: function (result) {
+                console.log(result);
+                if (result.success) {
+                    if ($.isFunction(callbackOnSuccess)) {
+                        callbackOnSuccess(result.data.active_package);
+                    }
+                } else {
+                    // @todo manage error
+                }
+			},
+            error: function (result) {
+                var result = result || new Object();
+                // @todo manage error
+            }
+		});
+	};
 	
 	/* Toogles the Bulk Action Check boxes */
 	Duplicator.Pack.SetDeleteAll = function() 
@@ -388,10 +491,13 @@ jQuery(document).ready(function($)
 		<?php $alert3->showAlert(); ?>
 	}
 
-	<?php if ($package_running) :?>
-		$('#pack-processing').show();
+<?php if ($package_running) : ?>
+    	$('#pack-processing').show();
+<?php endif;
 
-	<?php endif; ?>
+if ($active_package_present) :?>
+    Duplicator.Pack.GetActivePackageInfo(Duplicator.Pack.ActivePackageInfo);
+<?php endif; ?>
 	
 });
 </script>
